+++ ios/sample/NotificationService/NotificationService.swift
@@ -7,7 +7,7 @@
 //
 
 import UserNotifications
-//import SendbirdChatSDK
+// import SendbirdChatSDK
 
 class NotificationService: UNNotificationServiceExtension {
     /// > NOTE: How to mark messages as delivered: